+++ cmd/export/main.go
@@ -26,7 +26,7 @@ func main() {
 
 	applicationDB := appdb.NewAppDB(config.GetConfig())
 	height := applicationDB.GetLastHeight()
-	currentState, err := state.NewState(height, ldb, nil, nil, 0, 100000)
+	currentState, err := state.NewState(height, ldb, nil, 0, 100000)
 	if err != nil {
 		panic(err)
 	}