+++ core/state/accounts/accounts.go
@@ -109,9 +109,6 @@ func (a *Accounts) getOrderedDirtyAccounts() []types.Address {
 
 func (a *Accounts) AddBalance(address types.Address, coin types.CoinSymbol, amount *big.Int) {
 	balance := a.GetBalance(address, coin)
-	if balance.Cmp(big.NewInt(0)) == 0 {
-		a.bus.Coins().AddOwnerAddress(coin, address)
-	}
 	a.SetBalance(address, coin, big.NewInt(0).Add(balance, amount))
 }
 
@@ -142,20 +139,11 @@ func (a *Accounts) GetBalance(address types.Address, coin types.CoinSymbol) *big
 
 func (a *Accounts) SubBalance(address types.Address, coin types.CoinSymbol, amount *big.Int) {
 	balance := big.NewInt(0).Sub(a.GetBalance(address, coin), amount)
-	if balance.Cmp(big.NewInt(0)) == 0 {
-		a.bus.Coins().RemoveOwnerAddress(coin, address)
-	}
 	a.SetBalance(address, coin, balance)
 }
 
 func (a *Accounts) SetBalance(address types.Address, coin types.CoinSymbol, amount *big.Int) {
-	if amount.Cmp(big.NewInt(0)) == 0 {
-		a.bus.Coins().RemoveOwnerAddress(coin, address)
-	}
 	account := a.getOrNew(address)
-	if a.GetBalance(address, coin).Cmp(big.NewInt(0)) == 0 {
-		a.bus.Coins().AddOwnerAddress(coin, address)
-	}
 	account.setBalance(coin, amount)
 }