+++ core/state/coins/coins.go
@@ -2,16 +2,13 @@ package coins
 
 import (
 	"bytes"
-	"encoding/binary"
 	"fmt"
 	"github.com/MinterTeam/minter-go-node/core/state/bus"
 	"github.com/MinterTeam/minter-go-node/core/types"
 	"github.com/MinterTeam/minter-go-node/rlp"
 	"github.com/MinterTeam/minter-go-node/tree"
-	"github.com/xujiajun/nutsdb"
 	"math/big"
 	"sort"
-	"strings"
 )
 
 const (
@@ -29,11 +26,10 @@ type Coins struct {
 
 	bus  *bus.Bus
 	iavl tree.Tree
-	db   *nutsdb.DB
 }
 
-func NewCoins(stateBus *bus.Bus, iavl tree.Tree, db *nutsdb.DB) (*Coins, error) {
-	coins := &Coins{bus: stateBus, iavl: iavl, db: db, list: map[types.CoinSymbol]*Model{}, dirty: map[types.CoinSymbol]struct{}{}}
+func NewCoins(stateBus *bus.Bus, iavl tree.Tree) (*Coins, error) {
+	coins := &Coins{bus: stateBus, iavl: iavl, list: map[types.CoinSymbol]*Model{}, dirty: map[types.CoinSymbol]struct{}{}}
 	coins.bus.SetCoins(NewBus(coins))
 
 	return coins, nil
@@ -44,15 +40,6 @@ func (c *Coins) Commit() error {
 	for _, symbol := range coins {
 		coin := c.list[symbol]
 		delete(c.dirty, symbol)
-
-		if coin.isDeleted {
-			c.iavl.Remove(append([]byte{mainPrefix}, symbol[:]...))
-
-			path := append([]byte{mainPrefix}, symbol[:]...)
-			path = append(path, infoPrefix)
-			c.iavl.Remove(path)
-			continue
-		}
 
 		if coin.IsDirty() {
 			data, err := rlp.EncodeToBytes(coin)
@@ -123,17 +110,6 @@ func (c *Coins) AddReserve(symbol types.CoinSymbol, amount *big.Int) {
 	c.get(symbol).AddReserve(amount)
 }
 
-func (c *Coins) Sanitize(symbol types.CoinSymbol) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-
-	coin := c.get(symbol)
-	if coin.IsToDelete() {
-		c.delete(coin.symbol)
-	}
-}
-
 func (c *Coins) Create(symbol types.CoinSymbol, name string, volume *big.Int, crr uint, reserve *big.Int, maxSupply *big.Int) {
 	coin := &Model{
 		CName:      name,
@@ -200,27 +176,6 @@ func (c *Coins) markDirty(symbol types.CoinSymbol) {
 	c.dirty[symbol] = struct{}{}
 }
 
-func (c *Coins) delete(symbol types.CoinSymbol) {
-	coin := c.GetCoin(symbol)
-	if coin.isDeleted {
-		return
-	}
-	coin.isDeleted = true
-
-	accounts, candidates, frozenfunds := c.getOwners(symbol)
-	for _, address := range accounts {
-		c.bus.Accounts().DeleteCoin(address, symbol)
-	}
-
-	for _, pubkey := range candidates {
-		c.bus.Candidates().DeleteCoin(pubkey, symbol)
-	}
-
-	for _, height := range frozenfunds {
-		c.bus.FrozenFunds().DeleteCoin(height, symbol)
-	}
-}
-
 func (c *Coins) getOrderedDirtyCoins() []types.CoinSymbol {
 	keys := make([]types.CoinSymbol, 0, len(c.dirty))
 	for k := range c.dirty {
@@ -232,139 +187,6 @@ func (c *Coins) getOrderedDirtyCoins() []types.CoinSymbol {
 	})
 
 	return keys
-}
-
-func (c *Coins) AddOwnerAddress(symbol types.CoinSymbol, address types.Address) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-	err := c.db.Update(func(tx *nutsdb.Tx) error {
-		return tx.SAdd(ownerAccountsIndexBucket, symbol.Bytes(), address.Bytes())
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Coins) RemoveOwnerAddress(symbol types.CoinSymbol, address types.Address) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-	err := c.db.Update(func(tx *nutsdb.Tx) error {
-		return tx.SRem(ownerAccountsIndexBucket, symbol.Bytes(), address.Bytes())
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Coins) AddOwnerCandidate(symbol types.CoinSymbol, pubkey types.Pubkey) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-	err := c.db.Update(func(tx *nutsdb.Tx) error {
-		return tx.SAdd(ownerCandidatesIndexBucket, symbol.Bytes(), pubkey.Bytes())
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Coins) RemoveOwnerCandidate(symbol types.CoinSymbol, pubkey types.Pubkey) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-	err := c.db.Update(func(tx *nutsdb.Tx) error {
-		return tx.SRem(ownerCandidatesIndexBucket, symbol.Bytes(), pubkey.Bytes())
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Coins) AddOwnerFrozenFund(symbol types.CoinSymbol, height uint64) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-	b := make([]byte, 8)
-	binary.LittleEndian.PutUint64(b, height)
-
-	err := c.db.Update(func(tx *nutsdb.Tx) error {
-		return tx.SAdd(ownerFrozenFundsIndexBucket, symbol.Bytes(), b)
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Coins) RemoveOwnerFrozenFund(symbol types.CoinSymbol, height uint64) {
-	if symbol.IsBaseCoin() {
-		return
-	}
-	b := make([]byte, 8)
-	binary.LittleEndian.PutUint64(b, height)
-
-	err := c.db.Update(func(tx *nutsdb.Tx) error {
-		return tx.SRem(ownerFrozenFundsIndexBucket, symbol.Bytes(), b)
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Coins) getOwners(symbol types.CoinSymbol) ([]types.Address, []types.Pubkey, []uint64) {
-	var addresses []types.Address
-	err := c.db.View(func(tx *nutsdb.Tx) error {
-		items, err := tx.SMembers(ownerAccountsIndexBucket, symbol.Bytes())
-		if err != nil {
-			return err
-		}
-
-		for _, item := range items {
-			addresses = append(addresses, types.BytesToAddress(item))
-		}
-
-		return nil
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	var pubkeys []types.Pubkey
-	err = c.db.View(func(tx *nutsdb.Tx) error {
-		items, err := tx.SMembers(ownerCandidatesIndexBucket, symbol.Bytes())
-		if err != nil {
-			return err
-		}
-
-		for _, item := range items {
-			pubkeys = append(pubkeys, types.BytesToPubkey(item))
-		}
-
-		return nil
-	})
-	if err != nil && !strings.Contains(err.Error(), "not found bucket") {
-		panic(err)
-	}
-
-	var frozenfunds []uint64
-	err = c.db.View(func(tx *nutsdb.Tx) error {
-		items, err := tx.SMembers(ownerFrozenFundsIndexBucket, symbol.Bytes())
-		if err != nil {
-			return err
-		}
-
-		for _, item := range items {
-			frozenfunds = append(frozenfunds, binary.LittleEndian.Uint64(item))
-		}
-
-		return nil
-	})
-	if err != nil && !strings.Contains(err.Error(), "not found bucket") {
-		panic(err)
-	}
-
-	return addresses, pubkeys, frozenfunds
 }
 
 func (c *Coins) Export(state *types.AppState) {