+++ core/transaction/executor.go
@@ -169,10 +169,6 @@ func RunTx(context *state.State,
 
 	response.GasPrice = tx.GasPrice
 
-	if !isCheck && response.Code == code.OK {
-		context.Coins.Sanitize(tx.GasCoin)
-	}
-
 	if tx.Type == TypeCreateCoin {
 		response.GasUsed = createCoinGas
 		response.GasWanted = createCoinGas